Update node-discovery.go
Was going though some of the issues and I remember that 63char was to long. I dont own a super card but was wondering if this code will work. We could technically add all of the gpus with long names to be able to rename them. In theory this is simple enough.

Signed-off-by: deathlessdd <deathlessdd@gmail.com>